+++ src/ListManager.hpp
@@ -1,5 +1,6 @@
 /*
     Created by wint0r
+    Modified by Me
 */
 
 #ifndef __LISTMANAGER_HPP_
@@ -12,62 +13,43 @@
 #include <map>
 #include <Geode/Geode.hpp>
 
-enum class RatingType {
-    Classic,
-    Platformer,
-    Pending
-};
-
 struct ListRating {
-    RatingType type;
-    bool isExtreme;
-    std::string tier;
-    int id;
-    std::string name;
+    int tier = 0;
+    int id = -1;
+    std::string name = "?";
 };
 
-const std::map<std::string, std::string> NLW_TO_GDDP = {
-    { "Beginner", "Platinum" },
-    { "Easy", "Platinum" },
-    { "Medium", "Sapphire" },
-    { "Hard", "Sapphire" },
-    { "Very Hard", "Sapphire" },
-    { "Insane", "Jade" },
-    { "Extreme", "Emerald" },
-    { "Remorseless", "Emerald" },
-    { "Relentless", "Ruby" },
-    { "Terrifying", "Ruby" },
-
-    // I KNOW THESE ARE NOT NLW TIERS i'm just lazy ok :sob:
-    { "Jade", "Jade" },
-    { "Ruby", "Ruby" },
-    { "Diamond", "Diamond" },
-    { "Onyx", "Onyx" },
-    { "Amethyst", "Amethyst" },
-    { "Azurite", "Azurite" },
-    { "Obsidian", "Obsidian" }
-};
-
-const std::map<std::string, std::string> IDS_TO_GDDP = {
-    { "Beginner", "Gold" },
-    { "Easy", "Gold" },
-    { "Medium", "Gold" },
-    { "Hard", "Amber" },
-    { "Very Hard", "Amber" },
-    { "Insane", "Amber" },
-    { "Extreme", "Amber" },
-    { "Remorseless", "Amber" },
-    { "Relentless", "Amber" },
-    { "Terrifying", "Amber" },
+const std::map<int, std::string> GDDL_TO_GDDP = { //These ratings are based on the highest GDDL Tier in each GDDP Tier
+    { 14, "Gold" }, //not sure if there's any tier 14 insanes but might as well check anyways
+    { 15, "Gold" },
+    { 16, "Gold" },
+    { 17, "Gold" },
+    { 18, "Amber" },
+    { 19, "Amber" },
+    { 20, "Amber" },
+    { 21, "Platinum" },
+    { 22, "Platinum" },
+    { 23, "Sapphire" },
+    { 24, "Sapphire" },
+    { 25, "Sapphire" },
+    { 26, "Jade" },
+    { 27, "Jade" },
+    { 28, "Emerald" },
+    { 29, "Emerald" },
+    { 30, "Ruby" },
+    { 31, "Ruby" },
+    { 32, "Diamond" },
+    { 33, "Diamond" },
+    { 34, "Onyx" },
+    { 35, "Amethyst" }
 };
 
 struct ListManager {
     static std::vector<ListRating> ratings;
-    static bool fetchedNLWRatings, fetchedIDSRatings, fetchedPointercrateRatings;
+    static bool fetchedGDDLRatings;
 
     static void init();
-    static void parseResponse(matjson::Value val, bool isExtreme);
-    static void parsePointercrateResponse(matjson::Value val_part1, matjson::Value val_part2);
+    static void parseResponse(matjson::Value val);
     //static void throwError(std::string message);
     static std::optional<ListRating> getRating(int levelID);
     static std::string getSpriteName(GJGameLevel* level);